+++ include/gridtools/common/integral_constant.hpp
@@ -15,6 +15,25 @@
 #include "host_device.hpp"
 
 namespace gridtools {
+
+    // This predicate checks if the the class has `std::integral_constant` as a public base.
+    // Note that it is not the same as `class is an instantiation of gridtools::integral_constant`.
+    // Also it is not the same as `class has `integral nested value_type type and value static member`.
+    template <class, class = void>
+    struct is_integral_constant : std::false_type {};
+
+    template <class T>
+    struct is_integral_constant<T,
+        std::enable_if_t<std::is_base_of<std::integral_constant<typename T::value_type, T::value>, T>::value>>
+        : std::true_type {};
+
+    template <class T, int Val, class = void>
+    struct is_integral_constant_of : std::false_type {};
+
+    template <class T, int Val>
+    struct is_integral_constant_of<T, Val, std::enable_if_t<is_integral_constant<T>::value && T() == Val>>
+        : std::true_type {};
+
     template <class T, T V>
     struct integral_constant : std::integral_constant<T, V> {
         using type = integral_constant;
